refactor(skills): type/name/code flow clarifications
