+++ src/plugins/skills/src/traits/visualize_item.rs
